src/main/java/com/demotwo/sell/service/impl/ProductCategoryServiceimpl.java
package com.demotwo.sell.service.impl;

import com.demotwo.sell.dao.ProductCategoryRepository;
import com.demotwo.sell.entity.ProductCategory;
import com.demotwo.sell.service.ProductCategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

/**
 * @Author: liudongyang
 * @Date: 2018/9/15 1:58
 * @Desc:
 */
@Service
public class ProductCategoryServiceimpl implements ProductCategoryService {

    @Autowired
    private ProductCategoryRepository productCategoryRepository;

    @Override
    public ProductCategory findByCategory(Integer categoryId) {
        return productCategoryRepository.findOne(categoryId);
    }

    @Override
    public List<ProductCategory> findByAll() {
        return productCategoryRepository.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> categoryType) {
        return productCategoryRepository.findByCategoryTypeIn(categoryType);
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return productCategoryRepository.save(productCategory);
    }
}
